+++ internal/decoder/decoder.go
@@ -225,9 +225,11 @@ func (d *Decoder) decodeVersion(
 ) ([]flow.Record, *decodeError) {
 	switch version {
 	case flow.VersionNetFlowV5:
-		return decodeNetFlowV5(exporter, payload, dst)
+		issue := func(reason string) { es.countError(flow.VersionNetFlowV5, reason) }
+		return d.decodeNetFlowV5(exporter, payload, dst, issue)
 	case flow.VersionNetFlowV8:
-		return decodeNetFlowV8(exporter, payload, dst)
+		issue := func(reason string) { es.countError(flow.VersionNetFlowV8, reason) }
+		return d.decodeNetFlowV8(exporter, payload, dst, issue)
 	case flow.VersionNetFlowV9:
 		issue := func(reason string) { es.countError(flow.VersionNetFlowV9, reason) }
 		return d.decodeNetFlowV9(exporter, payload, dst, issue)